requirements.txt
# Core Dependencies
requests
python-dotenv
rich
openai
google-genai

# Screen Capture
mss
Pillow

# Speech to Text
SpeechRecognition
PyAudio

# Text to Speech
pyttsx3

# Hotkey Listener
pynput

# Web Demo (optional)
gradio

# HTTP utilities
urllib3
